Fix Scrambling Measurements
Changed how scrambling is calculated considering double-substituted species. measureScrambling now accepts a second input for R15a, R15b and R17 called "doubles" that are specific to the double-substituted species. Not using this input will default to using the same R values for double-substitution that was specified by the sample.